+++ hubdle-gradle-plugin/test/resources/kotlin/jvm/features/molecule/main/kotlin/com/javiersc/kotlin/jvm/molecule/FooEffect.kt
@@ -1,0 +1,6 @@
+package com.javiersc.kotlin.jvm.molecule
+
+enum class FooEffect {
+    A,
+    B,
+}