+++ hubdle-gradle-plugin/test/resources/kotlin/jvm/features/molecule/main/kotlin/com/javiersc/kotlin/jvm/molecule/FooState.kt
@@ -1,0 +1,3 @@
+package com.javiersc.kotlin.jvm.molecule
+
+data class FooState(val name: String, val counter: Int)